+++ lib/pages/forgot.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:thdapp/components/buttons/GradBox.dart';
 import 'custom_widgets.dart';
 import '../components/background_shapes/CurvedBottom.dart';
 import 'login.dart';
