+++ lib/pages/project_submission.dart
@@ -1,5 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:thdapp/api.dart';
+import 'package:thdapp/components/background_shapes/CurvedTop.dart';
+import 'package:thdapp/components/buttons/GradBox.dart';
+import 'package:thdapp/components/buttons/SolidButton.dart';
+import 'package:thdapp/components/topbar/TopBar.dart';
 import 'custom_widgets.dart';
 import 'enter_prizes.dart';
 import 'package:shared_preferences/shared_preferences.dart';
