+++ lib/pages/sponsors.dart
@@ -1,4 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:thdapp/components/background_shapes/CurvedTop.dart';
+import 'package:thdapp/components/buttons/GradBox.dart';
+import 'package:thdapp/components/buttons/SolidButton.dart';
+import 'package:thdapp/components/topbar/TopBar.dart';
 import 'package:thdapp/pages/bookmarks.dart';
 import 'custom_widgets.dart';
 import '../models/profile.dart';
